Bring Pipfile/requirements up to date for nhstravel library.
The Pipfile at the top level is used for the nhstravel library and for the unit tests.
THis brings the Pipfile to a place where at least all the tests run successfully.
Pipfile.lock and requirements.txt are also added to version control as these are current recommended best practice.
See:
https://github.com/pypa/pipenv/issues/598
https://stackoverflow.com/q/46278288