Update max_coverage_location to show non-included but overlapping LSOAs.
This is to fix https://github.com/nhs-pycom/nhs_time_of_travel/issues/54
Before this change, LSOAs which overlapped the circle, but their centroid was outside got left off.
We still don't want them to be included in the stats for covered population. However, it is useful to display them on the map,
So they are now shown in the map in red.

The algorithm was changed slightly. We use gpd.distance to find all relevant polygons for LSOAs. We do this in
British national grid space, partly because it is much quicker than converting the entire file to lat,long, and
partly because it makes distances easier as distances are already at 1m per unit.

A lot of now unused code was removed, and the file also formatted to standard formatting